fix: вынесла NpcControll в отдельный файл
